feat: add AbstractFFmpegOutputBuilder
