+++ src/main/java/net/bramp/ffmpeg/modelmapper/Mapper.java
@@ -2,6 +2,7 @@ package net.bramp.ffmpeg.modelmapper;
 
 import static net.bramp.ffmpeg.modelmapper.NotDefaultCondition.notDefault;
 
+import net.bramp.ffmpeg.builder.AbstractFFmpegOutputBuilder;
 import net.bramp.ffmpeg.builder.AbstractFFmpegStreamBuilder;
 import net.bramp.ffmpeg.builder.FFmpegOutputBuilder;
 import net.bramp.ffmpeg.options.AudioEncodingOptions;
@@ -72,23 +73,23 @@ public class Mapper {
     }
   }
 
-  public static <T extends AbstractFFmpegStreamBuilder<T>> void map(
-      MainEncodingOptions opts, AbstractFFmpegStreamBuilder<T> dest) {
+  public static <T extends AbstractFFmpegStreamBuilder<?>> void map(
+      MainEncodingOptions opts, T dest) {
     mapper.map(opts, dest);
   }
 
-  public static <T extends AbstractFFmpegStreamBuilder<T>> void map(
-      AudioEncodingOptions opts, AbstractFFmpegStreamBuilder<T> dest) {
+  public static  <T extends AbstractFFmpegStreamBuilder<?>> void map(
+      AudioEncodingOptions opts, T dest) {
     mapper.map(new AudioWrapper(opts), dest);
   }
 
-  public static <T extends AbstractFFmpegStreamBuilder<T>> void map(
-      VideoEncodingOptions opts, AbstractFFmpegStreamBuilder<T> dest) {
+  public static  <T extends AbstractFFmpegStreamBuilder<?>> void map(
+      VideoEncodingOptions opts, T dest) {
     mapper.map(new VideoWrapper(opts), dest);
   }
 
-  public static <T extends AbstractFFmpegStreamBuilder<T>> void map(
-      EncodingOptions opts, AbstractFFmpegStreamBuilder<T> dest) {
+  public static <T extends AbstractFFmpegStreamBuilder<?>> void map(
+      EncodingOptions opts, T dest) {
     map(opts.getMain(), dest);
 
     if (opts.getAudio().enabled) {
